Change casts of function pointers from int to long. This makes the compiler warnings go away, but the compiler is throwing away 32-bits as the long value is silently truncated to an int on alpha. But the program works, so that must not matter.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@35883 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ games/trek/computer.c
@@ -117,7 +117,7 @@ computer()
 	while (1)
 	{
 		r = getcodpar("\nRequest", Cputab);
-		switch ((int)r->value)
+		switch ((long)r->value)
 		{
 
 		  case 1:			/* star chart */
